Split parsers into separate modules
